requirements.txt
# 暴涨币种筛选器依赖包列表 / Surging Coin Screener Dependencies List
# 专门为市场扫描器和交易所API交互优化的依赖包 / Dependencies optimized for market scanner and exchange API interactions

# 核心数据处理库 / Core Data Processing Libraries
pandas>=1.5.0                 # 数据处理和分析 / Data processing and analysis
numpy>=1.21.0                 # 数值计算 / Numerical computing

# 异步HTTP客户端 / Async HTTP Client
aiohttp>=3.8.0                # 异步HTTP请求库 / Async HTTP request library
aiohttp-socks>=0.8.0          # SOCKS代理支持 / SOCKS proxy support
aiofiles>=0.8.0               # 异步文件操作 / Async file operations

# 配置文件处理 / Configuration File Processing
PyYAML>=6.0                   # YAML配置文件解析 / YAML config file parsing
python-dotenv>=1.0.0          # 环境变量文件支持 / Environment variables file support

# 日期时间处理 / Date Time Processing
python-dateutil>=2.8.0       # 日期时间工具 / Date time utilities

# 可选依赖：高性能数据处理 / Optional Dependencies: High Performance Data Processing
# 如果需要更好的性能，可以安装以下包 / Install the following packages for better performance
# numba>=0.56.0               # JIT编译器，加速数值计算 / JIT compiler for numerical computing
# ujson>=5.0.0                # 快速JSON解析 / Fast JSON parsing
# orjson>=3.8.0               # 更快的JSON处理 / Even faster JSON processing

# 可选依赖：通知功能 / Optional Dependencies: Notification Features
# requests>=2.28.0            # HTTP请求库（用于Webhook通知） / HTTP library for webhook notifications
# smtplib                     # 邮件发送（Python标准库） / Email sending (Python standard library)

# 开发和调试工具 / Development and Debugging Tools
# pytest>=7.0.0              # 测试框架 / Testing framework
# pytest-asyncio>=0.21.0     # 异步测试支持 / Async testing support
# black>=22.0.0               # 代码格式化 / Code formatting
# flake8>=4.0.0               # 代码检查 / Code linting

# 注意事项 / Notes:
# 1. 版本号使用 >= 以确保兼容性 / Version numbers use >= to ensure compatibility
# 2. 可选依赖已注释，按需取消注释安装 / Optional dependencies are commented, uncomment as needed
# 3. 如果遇到版本冲突，可以尝试降低版本要求 / If version conflicts occur, try lowering version requirements
# 4. 建议使用虚拟环境安装依赖 / Recommend using virtual environment for installation

# Python版本要求 / Python Version Requirements:
# Python >= 3.8 推荐 Python 3.9+ / Python >= 3.8, recommend Python 3.9+ 